refactor: make TodoInput.style.ts
`TodoInputForm` 컴포넌트를 만들었으니, 스타일도 분리함.
